chore(dev-console): Refactor and code cleanup
- Refactored OPC Base and Database
- And added some validation to add app form
- UI bug fix for Sidebar
- Fix for broken links in Overview and Sidebar
