+++ packages/developer-console-spa/package.json
@@ -41,7 +41,7 @@
     "yup": "0.32.9"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "PORT=5500 react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
