+++ packages/developer-console-spa/src/components/ConfigureOPCBase.module.css
